Audit-Armada-App refactoring
This update changes the auditd daemon start command to use
the "no fork" option.
Also the auditd start command will be wrapped in a shell
script that will be called in the container.
Other changes include renaming the armada app and the helm
charts to "auditd".
Renaming has been applied to all the applicable configuration.

Story: 2008849
Task: 42691

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: I8cf0659f820bc3968c69fa981cfc6342d5571a7c