Add auto_update for audit stx app.
This commit adds the auto update capability to the audit app.
This change will enable the audit app to be updated when a
new version is made available to the system.

Test cases:

PASS: build-pkgs -p stx-audit-helm

PASS: build-image

PASS: system application-apply auditd

PASS: In a system where auditd is in applied state, verify
      whether auto_update works by placing the new tarbal
      in "/usr/local/share/applications/helm/" and monitoring
      if sysinv automaticaly updates auditd to a new version.

Story: 2010628
Task: 48273

Change-Id: I1ffaba0fee9b34d237d252134e1680188cb36537
Signed-off-by: Alan Bandeira <Alan.PortelaBandeira@windriver.com>